Split up command packages by top-level subcommand
