+++ modules/cli/src/main/kotlin/org/coner/trailer/cli/command/config/ConfigDatabaseRemoveCommand.kt
@@ -1,4 +1,4 @@
-package org.coner.trailer.cli.command
+package org.coner.trailer.cli.command.config
 
 import com.github.ajalt.clikt.core.Abort
 import com.github.ajalt.clikt.core.CliktCommand
@@ -8,11 +8,16 @@ import com.github.ajalt.clikt.parameters.types.choice
 import org.coner.trailer.cli.io.ConfigurationService
 import org.coner.trailer.cli.io.DatabaseConfiguration
 
-class ConfigDatabaseSetDefaultCommand(
+class ConfigDatabaseRemoveCommand(
         private val config: ConfigurationService
 ) : CliktCommand(
-        name = "set-default",
-        help = "Set named database to default"
+        name = "remove",
+        help = """
+                Remove a database from the CLI app's config file.
+                
+                This is a non-destructive operation -- you can always add the database again. This will not affect the
+                database itself.
+            """.trimIndent()
 ) {
 
     val dbConfig: DatabaseConfiguration by option(names = *arrayOf("--name"))
@@ -21,9 +26,6 @@ class ConfigDatabaseSetDefaultCommand(
 
     override fun run() {
         if (dbConfig == config.noDatabase) throw Abort()
-        val asDefault = dbConfig.copy(
-                default = true
-        )
-        config.configureDatabase(asDefault)
+        config.removeDatabase(dbConfig)
     }
 }
