+++ modules/cli/src/main/kotlin/org/coner/trailer/cli/command/config/ConfigDatabaseSetDefaultCommand.kt
@@ -1,4 +1,4 @@
-package org.coner.trailer.cli.command
+package org.coner.trailer.cli.command.config
 
 import com.github.ajalt.clikt.core.Abort
 import com.github.ajalt.clikt.core.CliktCommand
@@ -8,16 +8,11 @@ import com.github.ajalt.clikt.parameters.types.choice
 import org.coner.trailer.cli.io.ConfigurationService
 import org.coner.trailer.cli.io.DatabaseConfiguration
 
-class ConfigDatabaseRemoveCommand(
+class ConfigDatabaseSetDefaultCommand(
         private val config: ConfigurationService
 ) : CliktCommand(
-        name = "remove",
-        help = """
-                Remove a database from the CLI app's config file.
-                
-                This is a non-destructive operation -- you can always add the database again. This will not affect the
-                database itself.
-            """.trimIndent()
+        name = "set-default",
+        help = "Set named database to default"
 ) {
 
     val dbConfig: DatabaseConfiguration by option(names = *arrayOf("--name"))
@@ -26,6 +21,9 @@ class ConfigDatabaseRemoveCommand(
 
     override fun run() {
         if (dbConfig == config.noDatabase) throw Abort()
-        config.removeDatabase(dbConfig)
+        val asDefault = dbConfig.copy(
+                default = true
+        )
+        config.configureDatabase(asDefault)
     }
 }
