+++ modules/cli/src/main/kotlin/org/coner/trailer/cli/di/CliktModule.kt
@@ -5,6 +5,8 @@ import com.github.ajalt.clikt.core.subcommands
 import com.github.ajalt.clikt.output.CliktConsole
 import com.github.ajalt.clikt.output.defaultCliktConsole
 import org.coner.trailer.cli.command.*
+import org.coner.trailer.cli.command.config.*
+import org.coner.trailer.cli.command.participanteventresultpointscalculator.*
 import org.kodein.di.DI
 import org.kodein.di.bind
 import org.kodein.di.instance
