+++ modules/cli/src/test/kotlin/org/coner/trailer/cli/command/participanteventresultpointscalculator/ParticipantEventResultPointsCalculatorGetCommandTest.kt
@@ -1,4 +1,4 @@
-package org.coner.trailer.cli.command
+package org.coner.trailer.cli.command.participanteventresultpointscalculator
 
 import assertk.assertThat
 import assertk.assertions.isEqualTo
@@ -16,9 +16,9 @@ import org.kodein.di.DI
 import org.kodein.di.bind
 import org.kodein.di.instance
 
-class ParticipantEventResultPointsCalculatorAddCommandTest {
+class ParticipantEventResultPointsCalculatorGetCommandTest {
 
-    lateinit var command: ParticipantEventResultPointsCalculatorAddCommand
+    lateinit var command: ParticipantEventResultPointsCalculatorGetCommand
 
     @MockK
     lateinit var service: ParticipantEventResultPointsCalculatorService
@@ -35,40 +35,44 @@ class ParticipantEventResultPointsCalculatorAddCommandTest {
     }
 
     @Test
-    fun `It should create calculator`() {
+    fun `It should get calculator by id`() {
         val calculator = TestParticipantEventResultPointsCalculators.lsccGroupingCalculator
-        every { service.hasNewName(calculator.name) } returns true
-        every { service.create(eq(calculator)) } answers { Unit }
-        val viewRenders = "created ${calculator.id}"
-        every { view.render(eq(calculator)) } returns viewRenders
+        every { service.findById(calculator.id) } returns calculator
+        val viewRenders = "rendered ${calculator.name}"
+        every { view.render(calculator) } returns viewRenders
 
-        command.parse(arrayOf(
-                "--id", calculator.id.toString(),
-                "--name", calculator.name,
-                "--position-to-points", "1", "9",
-                "--position-to-points", "2", "6",
-                "--position-to-points", "3", "4",
-                "--position-to-points", "4", "3",
-                "--position-to-points", "5", "2",
-                "--default-points", "1"
-        ))
+        command.parse(arrayOf("--id", calculator.id.toString()))
 
-        verifySequence {
-            service.hasNewName(calculator.name)
-            service.create(eq(calculator))
-            view.render(eq(calculator))
-        }
         assertThat(console.output).isEqualTo(viewRenders)
+        verifySequence {
+            service.findById(calculator.id)
+            view.render(calculator)
+        }
+    }
+
+    @Test
+    fun `It should get calculator by name`() {
+        val calculator = TestParticipantEventResultPointsCalculators.lsccGroupingCalculator
+        every { service.findByName(calculator.name) } returns calculator
+        val viewRenders = "rendered ${calculator.id}"
+        every { view.render(calculator) } returns viewRenders
+
+        command.parse(arrayOf("--name", calculator.name))
+
+        assertThat(console.output).isEqualTo(viewRenders)
+        verifySequence {
+            service.findByName(calculator.name)
+            view.render(calculator)
+        }
     }
 }
 
-
-private fun ParticipantEventResultPointsCalculatorAddCommandTest.arrangeCommand() {
+private fun ParticipantEventResultPointsCalculatorGetCommandTest.arrangeCommand() {
     val di = DI {
         bind<ParticipantEventResultPointsCalculatorService>() with instance(service)
         bind<ParticipantEventResultPointsCalculatorView>() with instance(view)
     }
-    command = ParticipantEventResultPointsCalculatorAddCommand(
+    command = ParticipantEventResultPointsCalculatorGetCommand(
             di = di,
             useConsole = console
     )
